Fix about page and split `index.js` into backend helper files
