+++ ui/arduino/views/about.css
@@ -1,0 +1,84 @@
+body,
+html {
+  width: 100%;
+  height: 100%;
+  -webkit-user-select: none;
+  user-select: none;
+  -webkit-app-region: drag;
+}
+
+body {
+  margin: 0;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  color: #333;
+  background-color: #eee;
+  font-size: 12px;
+  font-family: 'Helvetica', 'Arial', 'ãƒ’ãƒ©ã‚®ãƒŽè§’ã‚´ Pro W3', 'Hiragino Kaku Gothic Pro', 'ãƒ¡ã‚¤ãƒªã‚ª', Meiryo, 'ï¼­ï¼³ ï¼°ã‚´ã‚·ãƒƒã‚¯', 'MS PGothic', sans-serif;
+}
+
+.logo {
+  width: 80%;
+  -webkit-user-select: none;
+  user-select: none;
+}
+.logo img{
+  width: 100%;
+  height: auto;
+}
+
+.title,
+.copyright,
+.description {
+  margin: 0.2em;
+}
+
+.clickable {
+  cursor: pointer;
+}
+
+.description {
+  text-align: left;
+  margin: 2em;
+
+}
+
+.versions {
+  border-collapse: collapse;
+  margin-top: 1em;
+}
+
+.copyright,
+.versions {
+  color: #999;
+}
+
+.buttons {
+  margin-bottom: 1em;
+  text-align: center;
+}
+
+.buttons button {
+  margin-top: 1em;
+  width: 100px;
+  height: 24px;
+}
+
+.link {
+  cursor: pointer;
+  color: #80a0c2;
+}
+
+.bug-report-link {
+  position: absolute;
+  right: 1em;
+  bottom: 1em;
+}
+
+.clickable,
+.bug-report-link,
+.buttons button {
+  -webkit-app-region: no-drag;
+}
